+++ src/PsDac/Service/InstallPackageCommand.cs
@@ -24,15 +24,16 @@ namespace PsDac
         public string DatabaseName { get; set; }
 
         [Parameter()]
+        public SwitchParameter UpgradeExisting { get; set; }
+
+        [Parameter()]
         public ObjectType[] ExcludeObjectTypes { get; set; }
 
-        protected override void ProcessRecord()
+        protected override void AsyncProcessRecord()
         {
-            base.ProcessRecord();
-
             var options = new DacDeployOptions();
             options.ExcludeObjectTypes = ExcludeObjectTypes;
-            Service.Deploy(package: Package, targetDatabaseName: DatabaseName, options: options);
+            Service.Deploy(package: Package, targetDatabaseName: DatabaseName, upgradeExisting: UpgradeExisting.IsPresent, options: options);
         }
     }
 }